remove Pillow dependency
