+++ packages/pymedphys_analysis/src/pymedphys_analysis/_install_requires.py
@@ -1,7 +1,6 @@
 install_requires = [
     'pymedphys_coordsandscales',
     'pymedphys_utilities',
-    'Pillow',
     'scipy',
     'numpy < 1.16, >= 1.12',
     'pandas',
