+++ packages/pymedphys_labs/tests/paulking/film/data/FilmCalib_EBT_horz_strip.png
