+++ packages/pymedphys_labs/tests/paulking/film/data/FilmCalib_EBT_vert_strip.png
